refactor(bitwise.unit-test.ts): reorganize code
